Desktop/Health_Sympom_Analyser_PLP/requirements.txt
Flask
python-dotenv
transformers
torch
soundfile
huggingface_hub
